# storage/api-docs/api-docs.json
{
    "openapi": "3.0.0",
    "info": {
        "title": "Painting Mistery API",
        "description": "Documentación de la API",
        "version": "1.0.0"
    },
    "servers": [
        {
            "url": "http://127.0.0.1:8000/api",
            "description": "Servidor local"
        }
    ],
    "paths": {
...
        "/detalle-venta-curso": {
            "get": {
                "tags": [
                    "Detalle Venta Curso"
                ],
                "summary": "Listar cursos vendidos",
                "operationId": "5d91f030bf5039e477ea22f0e61d99a7",
                "responses": {
                    "200": {
                        "description": "Lista"
                    }
                }
            },
            "post": {
                "tags": [
                    "Detalle Venta Curso"
                ],
                "summary": "Agregar curso a una venta",
                "operationId": "e8969e6ed4d857c60fbccf07f684e50d",
                "requestBody": {
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "required": [
                                    "venta_id",
                                    "curso_id"
                                ],
                                "properties": {
                                    "venta_id": {
                                        "type": "integer"
                                    },
                                    "curso_id": {
                                        "type": "integer"
                                    }
                                },
                                "type": "object"
                            }
                        }
                    }
                },
                "responses": {
                    "201": {
                        "description": "Agregado"
                    },
                    "422": {
                        "description": "Error de validación"
                    },
                    "404": {
                        "description": "No encontrado"
                    }
                }
            }
        },
        "/detalle-venta-curso/{id}": {
            "get": {
                "tags": [
                    "Detalle Venta Curso"
                ],
                "summary": "Obtener detalle",
                "operationId": "028d1eada65403f693f9a0d4ae9a0d55",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Encontrado"
                    },
                    "404": {
                        "description": "No encontrado"
                    }
                }
            },
            "put": {
                "tags": [
                    "Detalle Venta Curso"
                ],
                "summary": "Actualizar detalle",
                "operationId": "739a4a05277ac53e32ebe03c74f120bd",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        }
                    }
                ],
                "requestBody": {
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "properties": {
                                    "curso_id": {
                                        "type": "integer"
                                    }
                                },
                                "type": "object"
                            }
                        }
                    }
                },
                "responses": {
                    "200": {
                        "description": "Actualizado"
                    },
                    "404": {
                        "description": "No encontrado"
                    },
                    "422": {
                        "description": "Error de validación"
                    }
                }
            },
            "delete": {
                "tags": [
                    "Detalle Venta Curso"
                ],
                "summary": "Eliminar detalle",
                "operationId": "f45b2265b73ea9a7e3e8b14d44f533b4",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Eliminado"
                    },
                    "404": {
                        "description": "No encontrado"
                    }
                }
            }
        },
...
    },
...
}